Sort srcset dictionary on render.
HTML doesn't care what order the srcset alternatives are in,
but sorting (by anything consistent) improves testability.

It does alpha sort on size.description - this will calc
the description more times than is ideal, but the dict sizes
are always going to be small enough that this makes no difference
in practice.